使用优惠券与不用优惠券费用查询后端完成 (#47)

+++ shared-trad-data-serv/shared-trad-data-api/src/main/java/cn/suparking/data/api/beans/ProjectConfig.java
@@ -15,7 +15,7 @@ public class ProjectConfig {
     private Integer txTTL;
 
     // 车位重复上报最小时间
-    private Long minIntervalForDupPark;
+    private Integer minIntervalForDupPark;
 
     // 付费免费时间
     private Integer prepayFreeMinutes;
